fix alignment issues in desktop, tablet and mobile view
adjusted alignment of properties focused on desktop initial aligment adjust the width of properties to help simplify styling for mobile responsiveness. adding play button within a div to make a block element. adjust min size of features in key benefits to display as 2 by 2 in tablet mode. adjusted the wing width in wb list to fit tablet viewport and taking off min width because it was unnecessary. adjust the pricing section to be centered on page and fit left half of page in desktop view. adjust the alignment of right side of hubspot section. and center the bullet titles in ref to bullet in list.